requirements.txt
mlflow==2.10.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4
apache-airflow==2.8.1
